Support scrobbling a show from the schedule
Next: support entering a custom date for scrobbles, which will be
useful if not scrobbling a show on the same date it aired.

I don't plan to support scrobbling shows from previous weeks,
though. Somebody else could implement this if desired, or add an
option to simply paste the show page's URL and get rows easily
(same process as for the homepage) that way.