Support shows with apostrophes in the name
Does not, and likely will not, support shows with BOTH a double
quote and an apostrophe in the name.

Not planning to add more features at this time -- this has been
a fun project, and this tool now supports all of my needs :-)

Maybe a little thing would be to print all of the song titles and
artists before prompting whether to scrobble all of them. Perhaps
even based on an optional verbose option. But the tool seems to be
getting the songs correct pretty reliably, and the pick menu is
still pretty convenient, so I don't see a pressing need for that
feature.

If someone would like to scrobble shows from previous weeks, by
all means add a PR and I will review it, but that's not a feature
I particularly want or need.